+++ src/components/sidebar/view/subcomponents/SidebarHeader.tsx
@@ -7,8 +7,6 @@ import { cn } from '../../../../lib/utils';
 import { useBranding } from '../../../../contexts/BrandingContext';
 import type { SidebarSearchMode } from '../../types/types';
 
-import GitHubStarBadge from './GitHubStarBadge';
-
 const MOD_KEY =
   typeof navigator !== 'undefined' && /Mac|iPhone|iPad/.test(navigator.platform) ? '⌘' : 'Ctrl';
 
@@ -133,8 +131,6 @@ export default function SidebarHeader({
             </Button>
           </div>
         </div>
-
-        <GitHubStarBadge />
 
         {/* Search bar */}
         {showSearchTools && (